+++ Plugins/mapsplat/style_converter.py
@@ -17,7 +17,7 @@ Supported style properties:
 - Multiple symbol layers
 """
 
-__version__ = "0.2.2"
+__version__ = "0.3.0"
 
 from qgis.core import (
     QgsVectorLayer,